Replace fly.io's TCP check with HTTP
Bandit logs an error when a client connects but doesn't make an HTTP
request. That means it was logging an error everytime fly.io did a TCP
check against the service without making an HTTP request.

Fixed by switching to an HTTP CHECK per https://community.fly.io/t/deploying-phoenix-with-the-bandit-adapter-on-fly-io/9327