Switch fly's check to an asset
Asset requests aren't logged because they are handled and returned
earlier in the plug stack. That means the check won't exercise a FULL
request to the app but it will at least confirm the app is running
enough so that plug knew to serve a static file so I'll take it.