Switch from Ash to Ecto for the recent comparisons
Ash seemed like it would work fine, but it made the app hit the out of
memory killer since each instance only has 256MB.

I could certainly bump up the memory but that seems excessive for
getting a pubsub view of recent comparisons.

One issue with this approach is that each instance should get the
broadcast recent comparions without issue, but when seeding from the
database they'll only get their local instance's recent comparisons.

To address that I'll either need to setup LiteFS or subscribe to the
updates and write the ones from non-local regions to the database.